requirements/local.txt
-r tests.txt

django-debug-toolbar==3.5.0
ipython==8.4.0
pre-commit==2.20.0
